requirements.txt
APScheduler==3.10.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorthief==0.2.1
idna==3.4
nider==0.5.0
oauthlib==3.2.2
Pillow==5.3.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tinydb==4.7.1
tweepy==4.13.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
